Quieted the expected, but confusing, exception output from some unittests.
